[tst-new-tabs-first] Listen 'new-tab-processed' TST event
Rather than listening to 'browser.tabs.onCreated', we listen to the TST
'event new-tab-processed'. Both TST and new-tabs-first listen to
'browser.tabs.onCreated' but it is undefined as to which listener will
fire first. new-tabs-first expects that TST's listener has already
processed the event so if new-tabs-first's listener fires first, the
extension will break.

By listening to 'new-tab-processed' event, we guarantee that TST will
have already processed the new tab.